Deny attempts to add duplicate bridge topics
It is possible to configure duplicate bridge topics with the same
parameters, but duplicates serve no purpose.

This change drops new topics which match exactly the parameters of an
existing topic, and returns MOSQ_ERR_SUCCESS since we do in fact have a
bridge topic which fulfills the request.

This change is to support a subsequent change to allow dynamically
adding bridge topics at runtime.