Make bridged topics 'dynamic'
This change implements dynamic bridge topics, which are created at
runtime when a client subscribes to a topic on the local broker.
That subscription triggers the creation of a new topic on the bridge,
which is reference counted.

This change DOES NOT implement topic removal in response to a local
unsubscribe, nor reference counting decrements. It is a work in
progress.